Refactor scope instrumenters
